requirements.txt
annotated-types==0.7.0
anyio==4.6.0
argcomplete==3.5.1
bcrypt==4.2.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.0
h11==0.14.0
idna==3.10
motor==3.6.0
packaging==24.1
passlib==1.7.4
pipx==1.7.1
platformdirs==4.3.6
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.9.2
python-dotenv==1.0.1
python-multipart==0.0.12
setuptools==75.1.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
userpath==1.9.2
uvicorn==0.31.0
wheel==0.44.0
